Migrate controller and registry to internal/controller2
- Remove the legacy harness build tag
