+++ cmd/ax/internal/cliutil/cliutil.go
@@ -1,5 +1,3 @@
-//go:build !harness
-
 // Copyright 2026 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,90 +17,105 @@ package cliutil
 import (
 	"context"
 	"fmt"
-	"time"
+	"os"
 
-	"github.com/google/ax/internal/agent"
-	"github.com/google/ax/internal/config"
-	"github.com/google/ax/internal/controller"
-	"github.com/google/ax/internal/controller/executor"
-	"github.com/google/ax/internal/gemini"
+	"github.com/google/ax/internal/config2"
+	"github.com/google/ax/internal/controller2"
+	"github.com/google/ax/internal/controller2/eventlog"
+	"github.com/google/ax/internal/harness"
 )
 
+const antigravityHarnessID = "antigravity"
+
 // Controller is the active controller type for this build.
-type Controller = *controller.Controller
+type Controller = *controller2.Controller
 
 // ExecHandler is the handler type accepted by Controller.Exec.
-type ExecHandler = controller.ExecHandler
+type ExecHandler = controller2.ExecHandler
 
 // Config is the configuration type for this build.
-type Config = config.Config
+type Config = config2.Config
 
 // LoadFromFile loads configuration from a YAML file.
 func LoadFromFile(path string) (*Config, error) {
-	return config.LoadFromFile(path)
+	return config2.LoadFromFile(path)
 }
 
 // DefaultConfig returns a configuration with default values set.
 func DefaultConfig() *Config {
-	return config.DefaultConfig()
+	return config2.DefaultConfig()
 }
 
-// NewControllerFromConfig creates a new Controller instance based on the provided configuration.
-func NewControllerFromConfig(ctx context.Context, cfg *Config) (*controller.Controller, error) {
-	// Validate planner type early
-	switch cfg.Planner.Type {
-	case "gemini":
-		// valid
-	default:
-		return nil, fmt.Errorf("unknown planner type: %s", cfg.Planner.Type)
+// NewControllerFromConfig creates a controller2.Controller instance based on the provided configuration.
+func NewControllerFromConfig(ctx context.Context, cfg *Config) (*controller2.Controller, error) {
+	reg := controller2.NewRegistry()
+
+	// AX_SUBSTRATE selects how built-in harnesses run: locally (unset) or as
+	// substrate actors ("1").
+	substrateMode := os.Getenv("AX_SUBSTRATE") == "1"
+
+	// Built-in harnesses.
+	var defaultHarnessID string
+	var antigravityHarness harness.Harness
+	var err error
+	if !substrateMode {
+		address := cfg.Harnesses.Antigravity.Endpoint
+		if address == "" {
+			address = "127.0.0.1:50053"
+		}
+		antigravityHarness = harness.NewAntigravityHarness(address)
+	} else {
+		antigravityHarness, err = harness.NewSubstrateHarness(antigravityHarnessID, "", "", "", 80)
+		if err != nil {
+			return nil, fmt.Errorf("antigravity harness: %w", err)
+		}
+	}
+	if err := reg.RegisterHarness(antigravityHarnessID, antigravityHarness); err != nil {
+		return nil, fmt.Errorf("register antigravity harness: %w", err)
+	}
+	if cfg.Harnesses.Antigravity.Default {
+		defaultHarnessID = antigravityHarnessID
 	}
 
-	// Create event log builder
-	eventLogBuilder := func() (executor.EventLog, error) {
-		return executor.OpenSQLiteEventLog(cfg.EventLog.SQLiteConfig.Filename)
+	// Custom substrate harnesses.
+	if len(cfg.Harnesses.Substrate) > 0 && !substrateMode {
+		return nil, fmt.Errorf("custom substrate harnesses require AX_SUBSTRATE=1")
+	}
+	for _, sc := range cfg.Harnesses.Substrate {
+		h, err := sc.NewHarness("")
+		if err != nil {
+			return nil, fmt.Errorf("substrate harness %q: %w", sc.ID, err)
+		}
+		if err := reg.RegisterHarness(sc.ID, h); err != nil {
+			return nil, fmt.Errorf("register substrate harness %q: %w", sc.ID, err)
+		}
+		if sc.Default {
+			defaultHarnessID = sc.ID
+		}
 	}
 
-	// Create planner builder
-	plannerBuilder := func(ctx context.Context, r *controller.Registry) (agent.Agent, error) {
-		switch cfg.Planner.Type {
-		case "gemini":
-			timeout, err := time.ParseDuration(cfg.Planner.Gemini.Timeout)
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse duration: %v", err)
+	// Register the configured default harness.
+	if defaultHarnessID != "" {
+		h, err := reg.Harness(defaultHarnessID)
+		if err != nil {
+			return nil, fmt.Errorf("default harness %q not found", defaultHarnessID)
+		}
+		if err := reg.RegisterHarness("", h); err != nil {
+			return nil, fmt.Errorf("register default harness %q: %w", defaultHarnessID, err)
+		}
+	}
+
+	return controller2.New(ctx, controller2.Config{
+		Registry: reg,
+		EventLogBuilder: func() (eventlog.EventLog, error) {
+			if cfg.EventLog.PostgresConfig.DSN != "" {
+				dsn := os.ExpandEnv(cfg.EventLog.PostgresConfig.DSN)
+				if dsn == "" {
+					return nil, fmt.Errorf("eventlog: postgres dsn %q expanded to empty", cfg.EventLog.PostgresConfig.DSN)
+				}
+				return eventlog.OpenPostgresEventLog(dsn)
 			}
-			return gemini.NewGeminiPlannerAgent(ctx, r, gemini.GeminiPlannerConfig{
-				GeminiConfig: &config.GeminiConfig{
-					Model:        cfg.Planner.Gemini.Model,
-					MaxTokens:    cfg.Planner.Gemini.MaxTokens,
-					Temperature:  cfg.Planner.Gemini.Temperature,
-					Timeout:      timeout,
-					SystemPrompt: cfg.Planner.Gemini.SystemPrompt,
-				},
-				SkillsDir: cfg.Planner.Gemini.SkillsDir,
-			})
-		default:
-			return nil, fmt.Errorf("unknown planner type: %s", cfg.Planner.Type)
-		}
-	}
-
-	// Build controller config
-	controllerConfig := controller.Config{
-		EventLogBuilder: eventLogBuilder,
-		PlannerBuilder:  plannerBuilder,
-	}
-
-	// Create controller
-	c, err := controller.New(ctx, controllerConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, agentCfg := range cfg.Registry.RemoteAgents {
-		if err := c.Registry().RegisterRemote(ctx, agentCfg); err != nil {
-			return nil, fmt.Errorf("failed to register remote agent %s: %w", agentCfg.ID, err)
-		}
-	}
-
-
-	return c, nil
+			return eventlog.OpenSQLiteEventLog(cfg.EventLog.SQLiteConfig.Filename)
+		},
+	})
 }
