--- internal/controller/executor/util.go
@@ -1,46 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package executor
-
-import (
-	"context"
-
-	"github.com/google/ax/internal/agent"
-	"github.com/google/ax/proto"
-)
-
-// agentFunc adapts a simple function into the agent.Agent interface.
-type agentFunc func(input []*proto.Message, tm agent.Executor, o agent.OutputHandler)
-
-func (f agentFunc) Connect(ctx context.Context, conversationID string, execID string, start *proto.AgentStart, tm agent.Executor, o agent.OutputHandler) error {
-	f(start.Messages, tm, o)
-	return nil
-}
-
-func (f agentFunc) Close() error { return nil }
-
-func AgentFunc(fn func(input []*proto.Message, tm agent.Executor, o agent.OutputHandler)) agent.Agent {
-	return agentFunc(fn)
-}
-
-// text is a helper that builds a plain-text Message.
-func text(role, s string) *proto.Message {
-	return &proto.Message{
-		Role: role,
-		Content: &proto.Content{
-			Type: &proto.Content_Text{Text: &proto.TextContent{Text: s}},
-		},
-	}
-}
